src/core/dragon_suite.py
"""
Main Dragon Suite class implementation
"""

from typing import Optional, Dict, Any
import yaml
from loguru import logger

class DragonSuite:
    """
    DragonSuite main class for managing automation workflows
    """
    
    def __init__(self, config_path: Optional[str] = None):
        """
        Initialize DragonSuite
        
        Args:
            config_path: Optional path to configuration file
        """
        self.config: Dict[str, Any] = {}
        if config_path:
            self.configure(config_path)
        
        logger.info("DragonSuite initialized")
    
    def configure(self, config_path: str) -> None:
        """
        Configure DragonSuite with settings from a YAML file
        
        Args:
            config_path: Path to configuration YAML file
        """
        try:
            with open(config_path, 'r') as f:
                self.config = yaml.safe_load(f)
            logger.info(f"Configuration loaded from {config_path}")
        except Exception as e:
            logger.error(f"Error loading configuration: {str(e)}")
            raise
    
    def run_automation(self, task_name: str, **kwargs) -> Any:
        """
        Run an automation task
        
        Args:
            task_name: Name of the task to run
            **kwargs: Additional arguments for the task
            
        Returns:
            Result of the automation task
        """
        logger.info(f"Running automation task: {task_name}")
        try:
            # TODO: Implement task execution logic
            pass
        except Exception as e:
            logger.error(f"Error running task {task_name}: {str(e)}")
            raise
            
    def add_integration(self, name: str, integration: Any) -> None:
        """
        Add a new integration to DragonSuite
        
        Args:
            name: Name of the integration
            integration: Integration instance
        """
        # TODO: Implement integration management
        pass
        
    def get_status(self) -> Dict[str, Any]:
        """
        Get current status of DragonSuite
        
        Returns:
            Dictionary containing status information
        """
        return {
            "status": "running",
            "config_loaded": bool(self.config),
            # Add more status information as needed
        }
