requirements.txt
# Core dependencies
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
